requirements.txt
pillow==10.2.0
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
controlnet_aux
PEFT
opencv-python
scipy==1.12.0
scikit-learn==1.3.0
opencv-python==4.8.0.74
albumentations==0.5.2
omegaconf==2.3.0
pytorch-lightning==2.0.6
kornia==0.5.0
numpy==1.23.5
fastapi==0.100.1
uvicorn==0.23.2
ultralytics==8.1.34
wget
mediapipe
runpod==.6.2 